Removed `devtools_test` dependencies on `devtools_app` test files
- Moved shared mocks/fakes in `devtools_app` into `devtools_test`
- Moved `waitFor(...)` helper function into `devtools_test`'s utils.dart
