+++ packages/devtools_app/test/layout_explorer/flex/flex_test.dart
@@ -7,12 +7,11 @@ import 'dart:convert';
 import 'package:devtools_app/src/inspector/diagnostics_node.dart';
 import 'package:devtools_app/src/inspector/inspector_tree.dart';
 import 'package:devtools_app/src/inspector/layout_explorer/flex/flex.dart';
+import 'package:devtools_test/mocks.dart';
 import 'package:devtools_test/wrappers.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import '../../inspector_screen_test.dart';
 
 // TODO(albertusangga): Re-enable tests in this files
 // https://github.com/flutter/devtools/issues/1403
